fix(velvet): read a mutant under the configurations that light its line
MutantParseabilityTests and MutantDeclarationRemovalTests parsed with
CSharpParseOptions.Default, so a line inside an `#if` region no symbol lights is
trivia to them: no tokens there to be unparseable, and no designation there to
lose. 124 of this package's 3255 line-removal mutants sit on such lines and
passed both guards unread.

MutantLineRemovalTests already derived three configurations from the source for
exactly this reason. That reading moves to MutantParseReadings and all three
fixtures share it, rather than a third copy -- the two older ones ask a different
question of the same tree, so what they want is the parse.

Parseability sums the diagnostics across all three, since a mutant unparseable
under any configuration is one the campaign cannot compile. The declaration guard
takes the widest single one instead: it compares one parse against another, and
the same line has to be a designation in both.

Measured, which is what this issue asks for first: with the guards able to read
those 124, nothing fails. 426 generator tests pass. So this is a ratchet against
a future generator defect on those lines rather than one it found today.
